requirements.txt
streamlit
google-generativeai
I am good at Maths and  problem solving, what career should i choose